Implemented disabled buttons instead of invisible buttons
The css is going to be easier to manage if the buttons are always visible but enabled and disabled depending on the game state.  Changes required to tests, new test functions based on element status, styles to disable the buttons on initial load, and the main js to invoke the necessary status change.